requirements.txt
elasticsearch>=8.0.0,<9.0.0
python-dotenv>=1.0.0
